🌱 Architect pass 306: split god files + fix frontend leaks (#15650)
- Split prediction_worker.go (879L) into 3 focused files:
  - prediction_worker.go (324L): types, struct, lifecycle
  - prediction_analysis.go (504L): analysis, data gathering, parsing
  - prediction_broadcast.go (72L): WebSocket broadcast logic

- Split bridge.go (696L) into 2 focused files:
  - bridge.go (569L): types, lifecycle, tool invocations
  - bridge_parsers.go (133L): result parsing helpers

- StockMarketTicker.tsx: wrap StockRow in React.memo to prevent
  full list re-renders on ticker/search state changes; add cancelled
  flag to debounced search effect preventing stale setState

- RSSFeed.tsx: reset feedInitRef on unmount so remounts re-fetch
